cleanup: restructure views
