+++ rythmize/api/v1/views/__init__.py
@@ -3,5 +3,7 @@ from flask import Blueprint
 
 api_views = Blueprint('api_views', __name__, url_prefix='/api/v1/')
 
-from .clients import *
-from .user_auth import *
+from .auth.spotify import *
+from .auth.users import *
+from .spotify_crud import *
+from .users_crud import *
