+++ rythmize/api/v1/views/spotify_crud.py
@@ -1,58 +1,12 @@
 """clients auth routes."""
 import flask_praetorian
-from flask import jsonify, request, redirect
+from flask import jsonify, request
 from rythmize.api.v1.views import api_views
 
 from ....clients.spotify import SpotifyClient
 from ....extensions import db_manager
-from ....models.user import User, UserSchema
+from ....models.user import User
 
-""" 
-Not needed for now
-@api_views.route('auth/connect/spotify/')
-@flask_praetorian.auth_required
-def generate_spotify_callback():
-    user_id = flask_praetorian.current_user().id
-    user = User.query.get(user_id)
-    if user.spotify_keys.jwt_token:
-        return jsonify("User already connected."), 410
-    sclient = SpotifyClient(None, user)
-    return jsonify({'auth_url': sclient.callback_uri_auth()}), 200
-"""
-
-@api_views.route('auth/connect/spotify/callback/')
-def authenticate_callback():
-    # get params
-    user_id = request.args.get('state')
-    user = User.query.get(user_id)
-    if 'error' in request.args.keys():
-        return jsonify("Failed to authenticate."), 401
-    code = request.args.get('code')
-    sclient = SpotifyClient(code, user)
-    data = sclient.handle_auth()
-    # update user linked table
-    spotify_table = user.spotify_keys
-    spotify_table.jwt_token = data["token"]
-    spotify_table.refresh_token = data["refresh_token"]
-    spotify_table.expires_in = data["expires"]
-    db_manager.save()
-
-    return redirect("https://rythmize-frontend.herokuapp.com/dashboard")
-
-
-@api_views.route('auth/connect/spotify/status', methods=["GET"])
-@flask_praetorian.auth_required
-def spotify_status():
-    user_id = flask_praetorian.current_user().id
-    user = User.query.get(user_id)
-    sclient = SpotifyClient(None, user)
-    if auth := sclient.handle_auth():
-        if type(auth) == dict:
-            user.spotify_keys.jwt_token = auth["token"]
-            user.spotify_keys.expires_in = auth["expires"]
-            db_manager.save()
-        return jsonify('User connected'), 200
-    return jsonify({"url": sclient.callback_uri_auth()}), 401
 
 @api_views.route('clients/spotify/playlists/', methods=["GET"])
 @flask_praetorian.auth_required
@@ -60,13 +14,14 @@ def get_user_playlists():
     user_id = flask_praetorian.current_user().id
     user = User.query.get(user_id)
     sclient = SpotifyClient(None, user)
-    if data:= sclient.handle_auth():
+    if data := sclient.handle_auth():
         if type(data) == dict:
             user.spotify_keys.jwt_token = data["token"]
             user.spotify_keys.expires_in = data["expires"]
             db_manager.save()
         return jsonify(sclient.get_user_playlists()), 200
     return jsonify("user not authorized"), 401
+
 
 @api_views.route('clients/spotify/playlists/<playlist_id>/tracks', methods=["GET"])
 @flask_praetorian.auth_required
@@ -75,7 +30,7 @@ def get_playlist_tracks(playlist_id):
     user_id = flask_praetorian.current_user().id
     user = User.query.get(user_id)
     sclient = SpotifyClient(None, user)
-    if data:= sclient.handle_auth():
+    if data := sclient.handle_auth():
         if type(data) == dict:
             user.spotify_keys.jwt_token = data["token"]
             user.spotify_keys.expires_in = data["expires"]
@@ -96,7 +51,7 @@ def spotify_transfer():
     songs = req.get('tracks', None)
     if playlist_name and songs and type(songs) == list:
         sclient = SpotifyClient(None, user)
-        if data:= sclient.handle_auth():
+        if data := sclient.handle_auth():
             if type(data) == dict:
                 user.spotify_keys.jwt_token = data["token"]
                 user.spotify_keys.expires_in = data["expires"]
